Rename `cupyx.rsqrt` submodule
