+++ cupyx/__init__.py
@@ -1,5 +1,5 @@
 # "NOQA" to suppress flake8 warning
-from cupyx.rsqrt import rsqrt  # NOQA
+from cupyx._rsqrt import rsqrt  # NOQA
 from cupyx._runtime import get_runtime_info  # NOQA
 from cupyx._scatter import scatter_add  # NOQA
 from cupyx._scatter import scatter_max  # NOQA
